move local/xgboost_submitter to step/xgboost (#3063)
