+++ python/runtime/xgboost/model_save_test.py
@@ -19,7 +19,7 @@ import pandas as pd
 import xgboost
 from jpmml_evaluator import make_evaluator
 from jpmml_evaluator.pyjnius import PyJNIusBackend, jnius_configure_classpath
-from runtime.local.xgboost_submitter.save import save_model_to_local_file
+from runtime.step.xgboost.save import save_model_to_local_file
 
 # Configure JVM
 jnius_configure_classpath()
