+++ python/runtime/xgboost/train.py
@@ -15,11 +15,11 @@ import sys
 
 import six
 import xgboost as xgb
-from runtime.local.xgboost_submitter.save import save_model_to_local_file
 from runtime.model import collect_metadata
 from runtime.model import oss as pai_model_store
 from runtime.model import save_metadata
 from runtime.pai.pai_distributed import make_distributed_info_without_evaluator
+from runtime.step.xgboost.save import save_model_to_local_file
 from runtime.xgboost.dataset import xgb_dataset
 from runtime.xgboost.pai_rabit import PaiXGBoostTracker, PaiXGBoostWorker
